+++ CHANGES.md
@@ -1,4 +1,5 @@
 # ?.?.?
+- add: new chart "speed per track"
 - fix: broken search page and improve responsiveness
 - fix: custom fields: show display name for type instead of enum name on settings page
 - fix: month goal distance: inputs not prefilled if german is activated