+++ CHANGES.md
@@ -1,5 +1,7 @@
-# ?.?.?
-- add: add quick filter buttons for track types on track and month goal page
+# 1.15.0
+- add: add quick filter buttons for track types on track page
+- add: add quick filter buttons for track types on month goal page
+- add: add quick filter buttons for track types on month search page
 - add: add quick filter buttons for track types and years on map page
 - add: new chart "speed per track"
 - fix: broken search page and improve responsiveness